shopperarmy/requirements.txt
tastypie
djangorestframework
markdown
django-filter
